+++ src/components/input/decorator/decorator.js
@@ -3,6 +3,9 @@ export function decorator() {
   if (this.input.current.value) {
     className += ' ' + this.valueClass
   }
+  if (this.state.autofill) {
+    className += ' ' + this.autofillClass
+  }
   if (this.state.focus) {
     className += ' ' + this.focusClass
   }