+++ src/components/text/text.js
@@ -17,9 +17,7 @@ export default class ManagedText extends ManagedInput {
       onMouseOver={this.onMouseOverHandler}
       onMouseLeave={this.onMouseLeaveHandler}
       onClick={this.onClickHandler}>
-        <div
-        style={{ animation: this.state.animation }}
-        className={this.labelClass + this.state.decorator}>
+        <div className={this.labelClass + this.state.decorator}>
           {this.state.label}
         </div>
         <input